Added cascading shadows.

+++ ice-tea-engine-core/src/main/java/com/destrostudios/icetea/core/data/values/Matrix4fArrayUniformValue.java
@@ -18,7 +18,7 @@ public class Matrix4fArrayUniformValue extends UniformValue<Matrix4f[]> {
 
     @Override
     public int getSize() {
-        return value.length * 16 * Integer.BYTES;
+        return value.length * 16 * Float.BYTES;
     }
 
     @Override
@@ -26,13 +26,12 @@ public class Matrix4fArrayUniformValue extends UniformValue<Matrix4f[]> {
         int currentIndex = index;
         for (Matrix4f matrix4f : value) {
             matrix4f.get(currentIndex, buffer);
-            currentIndex += 16 * Integer.BYTES;
+            currentIndex += 16 * Float.BYTES;
         }
     }
 
     @Override
     public String getShaderDefinitionType() {
-        // TODO: Clarify if this is fine and called again when the array is changed/resized - It's needed to access it by index inside shaders. After clarifying, do the same result for IntArrayUniformValue
         return "mat4[" + value.length + "]";
     }
